+++ src/subscriber/src/main/java/com/ioteste/subscriber/model/TemperatureReading.java
@@ -16,7 +16,7 @@ public record TemperatureReading(
         String roomId,
         double tempCelsius,
         double tempFahrenheit,
-        double sourceTs,
+        long sourceTs,
         Instant receivedAt
 ) {
 }
